requirements.txt
Django
djangorestframework
google-cloud-storage
python-dotenv
django-storages
